+++ app/src/main/java/com/poptech/poptalk/provider/SpeakItemModel.java
@@ -104,7 +104,7 @@ public class SpeakItemModel implements BaseModel {
                         speakItem.setId(cursor.getLong(SpeakItemQuery.SPEAK_ITEM_ID));
                         speakItem.setAudioPath(cursor.getString(SpeakItemQuery.SPEAK_ITEM_AUDIO_PATH));
                         speakItem.setAudioMark(cursor.getString(SpeakItemQuery.SPEAK_ITEM_AUDIO_MARK));
-                        speakItem.setAudioDuration(cursor.getInt(SpeakItemQuery.SPEAK_ITEM_AUDIO_DURATION));
+                        speakItem.setAudioDuration(cursor.getLong(SpeakItemQuery.SPEAK_ITEM_AUDIO_DURATION));
                         speakItem.setPhotoPath(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_PATH));
                         speakItem.setDescription1(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_DESCRIPTION1));
                         speakItem.setDescription2(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_DESCRIPTION2));
@@ -149,7 +149,7 @@ public class SpeakItemModel implements BaseModel {
                             speakItem.setId(cursor.getLong(SpeakItemQuery.SPEAK_ITEM_ID));
                             speakItem.setAudioPath(cursor.getString(SpeakItemQuery.SPEAK_ITEM_AUDIO_PATH));
                             speakItem.setAudioMark(cursor.getString(SpeakItemQuery.SPEAK_ITEM_AUDIO_MARK));
-                            speakItem.setAudioDuration(cursor.getInt(SpeakItemQuery.SPEAK_ITEM_AUDIO_DURATION));
+                            speakItem.setAudioDuration(cursor.getLong(SpeakItemQuery.SPEAK_ITEM_AUDIO_DURATION));
                             speakItem.setPhotoPath(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_PATH));
                             speakItem.setDescription1(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_DESCRIPTION1));
                             speakItem.setDescription2(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_DESCRIPTION2));
@@ -196,7 +196,7 @@ public class SpeakItemModel implements BaseModel {
                             speakItem.setId(cursor.getLong(SpeakItemQuery.SPEAK_ITEM_ID));
                             speakItem.setAudioPath(cursor.getString(SpeakItemQuery.SPEAK_ITEM_AUDIO_PATH));
                             speakItem.setAudioMark(cursor.getString(SpeakItemQuery.SPEAK_ITEM_AUDIO_MARK));
-                            speakItem.setAudioDuration(cursor.getInt(SpeakItemQuery.SPEAK_ITEM_AUDIO_DURATION));
+                            speakItem.setAudioDuration(cursor.getLong(SpeakItemQuery.SPEAK_ITEM_AUDIO_DURATION));
                             speakItem.setPhotoPath(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_PATH));
                             speakItem.setDescription1(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_DESCRIPTION1));
                             speakItem.setDescription2(cursor.getString(SpeakItemQuery.SPEAK_ITEM_PHOTO_DESCRIPTION2));
